attempt fixing the login
