+++ app/src/main/java/ch/epfl/sweng/zuluzulu/CommunicationTag.java
@@ -21,7 +21,6 @@ public enum CommunicationTag {
     OPEN_EVENT_DETAIL_FRAGMENT,
     OPEN_ASSOCIATION_FRAGMENT,
     OPEN_ASSOCIATION_DETAIL_FRAGMENT,
-    OPENING_WEBVIEW,
 
     // Admin tags
     OPEN_CREATE_EVENT,
