+++ app/src/main/res/layout/fragment_login.xml
@@ -25,48 +25,26 @@
             app:layout_constraintTop_toTopOf="parent" />
     </android.support.constraint.ConstraintLayout>
 
-    <ScrollView
-        android:id="@+id/login_form"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
 
         <android.support.constraint.ConstraintLayout
+            android:id="@+id/login_form"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
 
-            <Button
-                android:id="@+id/sign_in_button"
-                style="?android:textAppearanceSmall"
-                android:layout_width="274dp"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="55dp"
-                android:layout_marginLeft="55dp"
-                android:layout_marginRight="55dp"
-                android:layout_marginStart="55dp"
-                android:text="@string/action_sign_in"
-                android:textStyle="bold"
+            <WebView
+                android:id="@+id/webview"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent" />
-
-            <TextView
-                android:layout_width="148dp"
-                android:layout_height="34dp"
-                android:layout_marginBottom="130dp"
-                android:layout_marginEnd="118dp"
-                android:layout_marginLeft="118dp"
-                android:layout_marginRight="118dp"
-                android:layout_marginStart="118dp"
-                android:layout_marginTop="178dp"
-                android:text="Press to login"
-                android:textSize="24sp"
-                app:layout_constraintBottom_toTopOf="@+id/sign_in_button"
-                app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
+                app:layout_constraintTop_toTopOf="parent"></WebView>
 
         </android.support.constraint.ConstraintLayout>
-    </ScrollView>
 
 </FrameLayout>
